Refactor GET all questions query into a function
